requirements/base.txt
django-braces==1.3.1
django-haystack==2.1.0
django-reversion==1.8.0
Django==1.6.1
easy-thumbnails
pillow==2.0
South==0.8.4
